refactor(jira): Moved generic Jira class to common package.
