+++ channel/src/main/java/com/synopsys/integration/alert/channel/jira2/server/JiraServerSearcher.java
@@ -10,7 +10,7 @@ package com.synopsys.integration.alert.channel.jira2.server;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import com.synopsys.integration.alert.channel.jira2.cloud.JiraIssueAlertPropertiesManager;
+import com.synopsys.integration.alert.channel.jira2.common.JiraIssueAlertPropertiesManager;
 import com.synopsys.integration.alert.channel.jira2.common.JiraSearcher;
 import com.synopsys.integration.alert.channel.jira2.common.model.JiraSearcherResponseModel;
 import com.synopsys.integration.exception.IntegrationException;
